+++ A1/src/Vertex.h
@@ -8,8 +8,8 @@
 struct Vertex {
     Position position;
     Color color;
-    Vertex(std::vector<float>& pos, std::vector<float>& color):position(pos), color(color) {}
-    Vertex(std::vector<float>& pos):position(pos), color() {}
+    Vertex(std::vector<int>& pos, std::vector<int>& color):position(pos), color(color) {}
+    Vertex(std::vector<int>& pos):position(pos), color() {}
 };
 
 #endif
